move function to get root of repository to utils package
